Fixed bug in "above ___" functions (and renamed)
Turns out that the internet isn't always right! Trying to check the block at playery-1  to see what the player is standing on will get you the incorrect block for anything that's not a full block in height.

Here's some examples to show why subtracting a very small number works. Say you're at height 16, standing on a block 1 in height. If you subtract 1, that checks what's at height 15 - the bllock below you. But you's still get the correct result if checking at 15.1, 15.5, or even 15.99999 - just so long as you don't go up to 16. So while -1 works, it only just BARELY works.

On the other hand, say you're on a carpet. Carpets are 1/16 in height, meaning you might be at y=15.0625. The ranges of y that give you the correct block START at 15. If you subtract 1, you'll be checking 14.0625 which is actually the block the carpet itself has been placed on top of (and not the carpet). So the solution is to subtract a number smaller than the height ofthe carpet (or if you want to be more general, just use a number smaller than the height of the shorted mc block. In this case I'm just using .01).

Note that this means we're only detecting what we're STANDING on. If you're at all eleveated. e.g. at teh very bottom of your jump, we'll no longer count you as on top of the block. But in this case that's probably behavior we want anyways.

Side note - also seeing interesting behavior when crouching off a side of a block. There's potentially some work I could do here, as you can 'cheat' the system by couching off the edge of one block indefinitely without triggering. Buuut honestly that's probably not a very big concern. If I do wnat to catch it though, I can probably tell the player is 'floating' this manner by seeing if they're sneaikng with an air block  - and then  figure out which block they're sneaking from by examining the fractional portion of their dimension (I think the max you can sneak off is around .3 of a block. So if they're on air and sneaking, I think x%1 < .3 means from the left, >.7 means from the right.