Fixed broken predicate AND-ing. host tag, and redefined flowers / buds
1. I wasn't negating the sub-terms properly, so the predicates were not actually doing the 'simulated AND' correctly
2. Predicate for 'host' was checking if the block was tagged a virus, not a host. oops.
3. Decided to adjust setup in order to take advantage of the cool look of crying obsidian, plus make it harder for a player to 'cure' infections.:

Now 'buds' are crying obsidian that touch purpur on any of its faces. And 'flowers' are carpets on top of buds that are on top of the viruses (meaning that it's a 'top' bud of the 6 surrounding faces on the virus).

End behavior should be a cool 'shell' of crying obsidian grown around virus blocks. Haven't decided if I want it to be more like an 'exoskeleton' for entire swathes or more like a cool patchwork effect, though I'm right now leaning towards the latter.
Can implement it either way by adjusting how crying obsidian appears - if it only grows to fill air spaces, then it'd be more like an exoskeleton. If I however instead make 'infections' a process where virus creates buds, and buds infect hosts to make more viruses, I'd potentially get cool efects. Like if you have one virus block being a 'core', it grows crying obsidian around that one block. Whioch in the end gives a really awesome looking 'pathwork' effect where players see mostly crying obsidian, and to clear infections need to mine through the shell to clear out the inner virus block. Though this might be rather obnoxious to clean up for the players.... Hmm. Something to think about

Might in the future want to make it consider any purple carpet on buds to be 'flowers' rather than requiring it to be the 'top' bud.  Should be pretty easy to do honestly if I just replace the current checks with one that looks to see if the below block passes the is_virus_bud predicate (rathjer than checking block tags)

Also tweaked readme since I don't think we can choose items from a list of tags when deciding what blocks to place. If the 'spawning'/infecting part might be hardcoded to place these blocks then just changing the tags won't be enough to make other types of blocks 'viral'.