Adding terracotta as flower to break up monotony of carpet + add 'value'
This means that our viral blocks provide some free terracotta for the players, make the terrain more 'interesting', and also cause more headaches by clogging up their inventory + not breaking when the below blocks are destroyed :)

Probably will want to have some ration like 10:1 carpet to terracotta when growing a bud. And then to choose randomly among the terracotta.
Also on that note, would be cool if buds have a small (1/50?) chance of growing another bud instead of a flower on them. That way we will rarely have some interesting weird 'trees' of buds grow out of the shells.